Mise a jour du service technicien pour permettre la synchronisation avec la table user lors de l'ajout, updated et deleted

+++ Services/ImplementationServices/ServiceTechnicien.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.EntityFrameworkCore;
+﻿using Microsoft.AspNetCore.Identity;
+using Microsoft.EntityFrameworkCore;
 using PMT.Data;
 using PMT.Models;
 using System;
@@ -11,6 +12,23 @@ namespace PMT.Services.ImplementationServices
     public class ServiceTechnicien : IServiceTechnicien
     {
         private readonly Db_Context context;
+        private readonly UserManager<User_App> userManager;
+        private readonly IUserValidator<User_App> userValidator;
+        private readonly IPasswordValidator<User_App> passwordValidator;
+        private readonly IPasswordHasher<User_App> passwordHasher;
+
+        public List<Erreur> erreurs = null;
+
+        public ServiceTechnicien(UserManager<User_App> userManager,
+            IUserValidator<User_App> userValidator,
+            IPasswordValidator<User_App> passwordValidator,
+            IPasswordHasher<User_App> passwordHasher)
+        {
+            this.userManager = userManager;
+            this.userValidator = userValidator;
+            this.passwordValidator = passwordValidator;
+            this.passwordHasher = passwordHasher;
+        }
 
         public ServiceTechnicien(Db_Context context)
         {
@@ -19,16 +37,40 @@ namespace PMT.Services.ImplementationServices
 
         public async Task<Technicien> Create(Technicien technicien)
         {
-            context.Techniciens.Add(technicien);
-            await context.SaveChangesAsync();
+            User_App user = new User_App
+            {
+                UserName = technicien.Username,
+                Email = technicien.Mail
+            };
+
+            IdentityResult result = await userManager.CreateAsync(user, technicien.Password);
+
+            if (result.Succeeded)
+            {
+                user = await userManager.FindByNameAsync(technicien.Username);
+                technicien.ID = user.Id;
+                context.Techniciens.Add(technicien);
+                await context.SaveChangesAsync();
+                return technicien;
+            }
             return technicien;
         }
 
-        public async Task Delete(string id)
+        public async Task<string> Delete(string id)
         {
-            var technicien = await context.Techniciens.FindAsync(id);
-            context.Techniciens.Remove(technicien);
-            await context.SaveChangesAsync();
+            User_App user = await userManager.FindByIdAsync(id);
+            if(user != null)
+            {
+                IdentityResult result = await userManager.DeleteAsync(user);
+                var technicien = await context.Techniciens.FindAsync(id);
+                context.Techniciens.Remove(technicien);
+                await context.SaveChangesAsync();
+                return "Suppression réussite";
+            }
+            else
+            {
+                return "Echec de suppression, l'utilisateur n'existe pas";
+            }
         }
 
         public async Task<IEnumerable<Technicien>> Get()
@@ -52,11 +94,37 @@ namespace PMT.Services.ImplementationServices
             return await context.Taches.AsNoTracking().Where(t => t.TechnicienID == idUser).ToListAsync();
         }
 
-        public async Task Update(Technicien technicien)
+        public async Task<string> Update(Technicien technicien, string id)
         {
-            context.Entry(technicien).State = EntityState.Modified;
-            await context.SaveChangesAsync();
+            User_App user = await userManager.FindByIdAsync(id);
+            if(user != null)
+            {
+                user.Email = technicien.Mail;
+                IdentityResult ValidEmail = await userValidator.ValidateAsync(userManager, user);
+                
+                IdentityResult validPass = null;
+                if (!string.IsNullOrEmpty(technicien.Password))
+                {
+                    validPass = await passwordValidator.ValidateAsync(userManager, user, technicien.Password);
+                    if (validPass.Succeeded)
+                    {
+                        user.PasswordHash = passwordHasher.HashPassword(user, technicien.Password);
+                    }
+                }
+
+                if((ValidEmail.Succeeded && validPass == null) || (ValidEmail.Succeeded && technicien.Password != string.Empty && validPass.Succeeded))
+                {
+                    IdentityResult result = await userManager.UpdateAsync(user);
+                    if (result.Succeeded)
+                    {
+
+                        context.Entry(technicien).State = EntityState.Modified;
+                        await context.SaveChangesAsync();
+                        return "Modification réussite";
+                    }
+                }
+            }
+            return "Echec de modification";
         }
     }
 }
-
